Restructure directories, prep for deployment
