+++ core/templates/core/index.html
