+++ core/templates/core/signup.html
